refactor: move recreateTables() to vite-plugin-db
